Moved BpsRenderer out of Indoor.h/cpp
