+++ src/Engine/Engine.cpp
@@ -1,6 +1,7 @@
 #include "Engine/Engine.h"
 
 #include "Engine/Events.h"
+#include "Engine/Graphics/Camera.h"
 #include "Engine/Graphics/DecalBuilder.h"
 #include "Engine/Graphics/DecorationList.h"
 #include "Engine/Graphics/IRender.h"
