+++ src/Engine/Graphics/ParticleEngine.cpp
@@ -1,5 +1,6 @@
 #include "Engine/Graphics/ParticleEngine.h"
 
+#include "Engine/Graphics/Camera.h"
 #include "Engine/LOD.h"
 #include "Engine/OurMath.h"
 #include "Engine/Time.h"
