+++ src/Engine/Graphics/Vis.cpp
@@ -10,6 +10,7 @@
 #include "Engine/OurMath.h"
 
 #include "Engine/Graphics/Level/Decoration.h"
+#include "Engine/Graphics/BspRenderer.h"
 #include "Engine/Graphics/Outdoor.h"
 #include "Engine/Graphics/Sprites.h"
 #include "Engine/Graphics/Viewport.h"
