+++ src/Engine/SpellFxRenderer.cpp
@@ -9,6 +9,7 @@
 #include "Engine/Time.h"
 #include "Engine/stru160.h"
 
+#include "Engine/Graphics/Camera.h"
 #include "Engine/Graphics/Indoor.h"
 #include "Engine/Graphics/LightmapBuilder.h"
 #include "Engine/Graphics/LightsStack.h"
