+++ src/GUI/UI/UIHouses.cpp
@@ -8,6 +8,7 @@
 #include "Engine/Awards.h"
 #include "Engine/Engine.h"
 #include "Engine/Events.h"
+#include "Engine/Graphics/Camera.h"
 #include "Engine/Graphics/Image.h"
 #include "Engine/Graphics/Level/Decoration.h"
 #include "Engine/Graphics/Outdoor.h"
